test: split PDF link target resolution cases
